+++ test.js
@@ -9,7 +9,7 @@ test('holiday_1', function(t){
 });
 
 test('holiday_2', function(t){
-  var w = moment('2010-03-25').holiday(['Turkey Day', 'Christmas']);
+  var w = moment('2010-03-25').holidays(['Turkey Day', 'Christmas']);
   var k = Object.keys(w);
   t.is(typeof w, 'object');
   t.is(k.length, 2);
@@ -23,9 +23,9 @@ test('holiday_3', function(t){
 });
 
 test('holiday_4', function(t){
-  var w = moment().holiday();
+  var w = moment().holidays();
   t.is(typeof w, 'object');
-  t.is(Object.keys(w).length, Object.keys(moment.fn.holidays).length);
+  t.is(Object.keys(w).length, Object.keys(moment.holidays).length);
 });
 
 test('holiday_5', function(t){
@@ -49,11 +49,39 @@ test('isHoliday_3', function(t){
   t.is(w, "Veteran's Day");
 });
 
+test('previousHoliday_1', function(t){
+  var w = moment('2002-06-15').previousHoliday(6);
+  t.is(w.constructor, Array);
+  t.is(w.length, 6);
+  t.is(w[4].isHoliday(), "Valentine's Day");
+});
+
+test('previousHoliday_2', function(t){
+  var w = moment('2012-02-06').previousHolidays(7, true);
+  t.is(w.constructor, Array);
+  t.is(w.length, 7);
+  t.is(w[2].isHoliday(true), "New Year's Eve");
+});
+
+test('nextHoliday_1', function(t){
+  var w = moment('2001-04-20').nextHolidays(5);
+  t.is(w.constructor, Array);
+  t.is(w.length, 5);
+  t.is(w[2].isHoliday(), "Father's Day");
+});
+
+test('nextHoliday_2', function(t){
+  var w = moment('2011-11-02').nextHoliday(8, true);
+  t.is(w.constructor, Array);
+  t.is(w.length, 8);
+  t.is(w[6].isHoliday(true), "New Year's Day");
+});
+
 test('holidaysBetween_1', function(t){
-  var w = moment('2011-11-01').holidaysBetween('2011-12-31');
-  t.is(typeof w, 'object');
-  t.is(Object.keys(w).length, 6);
-  t.is(w[Object.keys(w)[4]].isHoliday(), 'Christmas Day');
+  var w = moment('2011-11-01').holidaysBetween('2012-05-15');
+  t.is(w.constructor, Array);
+  t.is(w.length, 12);
+  t.is(w[11].isHoliday(), "Mother's Day");
 });
 
 test('holidaysBetween_2', function(t){
@@ -63,9 +91,9 @@ test('holidaysBetween_2', function(t){
 
 test('holidaysBetween_3', function(t){
   var w = moment('2011-11-01').holidaysBetween('2011-12-31', true);
-  t.is(typeof w, 'object');
-  t.is(Object.keys(w).length, 6);
-  t.true(w[Object.keys(w)[5]].isSame('2011-12-30', 'day'));
+  t.is(w.constructor, Array);
+  t.is(w.length, 6);
+  t.true(w[5].isSame('2011-12-30', 'day'));
 });
 
 test('modifyHolidays_set_1', function(t){