+++ CHANGELOG.md
@@ -5,8 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
-## 1.0.0 - TBA
+
+## 1.1.0 - 2017-07-17
+### Added
+- `previousHoliday` and `nextHoliday` functions.
+- `holidays` alias function for `holiday`.
+
+### Changed
+- Moved holidays object from `moment.fn.holidays` to `moment.holidays`.
+- Fixed bug where `holidaysBetween` function would only return holidays for the start year.
+- `holidaysBetween` function now returns an array of moment objects rather than an object.
+
+
+## 1.0.0 - 2017-07-16
 ### Added
 - Initial release.
 
 [Unreleased]: https://github.com/kodie/moment-holiday/compare/v1.0.0...HEAD
+[1.1.0]: https://github.com/kodie/moment-holiday/compare/v1.0.0...v1.1.0